refactor: move stores to related modules
