+++ packages/theme/modules/catalog/category/stores/category.ts
