+++ packages/theme/modules/checkout/stores/cart.ts
@@ -1,0 +1,14 @@
+import { defineStore } from 'pinia';
+import type { Cart } from '~/modules/GraphQL/types';
+
+interface CustomerState {
+  cart: Cart,
+}
+
+export const useCartStore = defineStore('cart', {
+  state: (): CustomerState => ({
+    cart: {
+      id: '', is_virtual: false, total_quantity: 0, shipping_addresses: [],
+    },
+  }),
+});
