+++ packages/theme/modules/customer/composables/useUser/index.ts
@@ -7,7 +7,7 @@ import {
 import type { Ref } from '@nuxtjs/composition-api';
 import mask from '~/composables/utils/mask';
 import { Logger } from '~/helpers/logger';
-import { useCustomerStore } from '~/stores/customer';
+import { useCustomerStore } from '~/modules/customer/stores/customer';
 import { useCart } from '~/modules/checkout/composables/useCart';
 import { generateUserData } from '~/modules/customer/helpers/generateUserData';
 import { Customer } from '~/modules/GraphQL/types';
