+++ packages/theme/modules/customer/stores/customer.ts
@@ -1,8 +1,7 @@
 import { defineStore } from 'pinia';
-import type { Cart, Customer } from '~/modules/GraphQL/types';
+import type { Customer } from '~/modules/GraphQL/types';
 
 interface CustomerState {
-  cart: Cart,
   user: Customer | null,
   isLoggedIn: boolean,
 }
@@ -10,9 +9,6 @@ interface CustomerState {
 export const useCustomerStore = defineStore('customer', {
   state: (): CustomerState => ({
     user: null,
-    cart: {
-      id: '', is_virtual: false, total_quantity: 0, shipping_addresses: [],
-    },
     isLoggedIn: false,
   }),
   actions: {
